Moved js files to src folder. Removed unused constants. Translate comments to english.
