+++ package.json
@@ -12,5 +12,11 @@
     "express": "^4.18.2",
     "puppeteer": "^21.3.0",
     "puppeteer-stream": "^3.0.7"
+  },
+  "devDependencies": {
+    "@types/body-parser": "^1.19.3",
+    "@types/express": "^4.17.18",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.2.2"
   }
 }
