+++ public/index.html
@@ -19,6 +19,7 @@
             padding: 20px;
             border-radius: 5px;
             box-shadow: 0 0 10px rgba(0, 0, 0, 0.2);
+            margin: 10px;
         }
 
         .form-container label {
@@ -31,7 +32,7 @@
         .form-container input[type="datetime-local"],
         .form-container input[type="password"],
         .form-container button {
-            width: 100%;
+            width: 90%;
             padding: 10px;
             margin-bottom: 10px;
             border: none;
@@ -59,7 +60,7 @@
         <input type="text" id="username" name="username">
 
         <label for="password">Пароль:</label>
-        <input type="password" id="password" name="password">
+        <input type="text" id="password" name="password">
 
         <button type="submit">Начать</button>
     </form>
